requirements.txt
sentence-transformers
ctransformers
langchain
streamlit
langchain_community
pymongo
pandas
langchain-huggingface
python-dotenv